feat: 完善

+++ repository/sys_role_permission.go
@@ -36,7 +36,7 @@ func (r *RolePermissionRepository) DeleteById(id uint) error {
 	return nil
 }
 
-func (r *RolePermissionRepository) UpdateById(id int64, PermissionKeys string) (*model.RolePermission, error) {
+func (r *RolePermissionRepository) UpdateById(id uint, PermissionKeys string) (*model.RolePermission, error) {
 	var ret = new(model.RolePermission)
 	data := &RolePermission{}
 	data.PermissionKeys = PermissionKeys
